+++ tests/attribute_value.rs
@@ -10,7 +10,6 @@ use svgparser::{
     PaintFallback,
     ElementId,
     Color,
-    Stream,
     Error,
     ErrorPos,
     ValueId,
@@ -20,8 +19,7 @@ macro_rules! test {
     ($name:ident, $aid:expr, $text:expr, $result:expr) => (
         #[test]
         fn $name() {
-            let mut s = Stream::from_str($text);
-            let v = AV::from_stream(ElementId::Rect, $aid, &mut s).unwrap();
+            let v = AV::from_str(ElementId::Rect, $aid, $text).unwrap();
             assert_eq!(v, $result);
         }
     )
@@ -31,8 +29,7 @@ macro_rules! test_err {
     ($name:ident, $aid:expr, $text:expr, $err:expr) => (
         #[test]
         fn $name() {
-            let mut s = Stream::from_str($text);
-            let v = AV::from_stream(ElementId::Rect, $aid, &mut s);
+            let v = AV::from_str(ElementId::Rect, $aid, $text);
             assert_eq!(v.unwrap_err(), $err);
         }
     )